Add the revision_create event
We have decided that we will follow all revision creation events, not
just page edits. This new scheme reflects that.

Differences from page_edit:
- renamed title to page_title
- renamed namespace to page_namespace
- renamed revision_id to rev_id
- renamed parent_revision_id to rev_parent_id
- renamed save_dt to rev_timestamp
- renamed summary to comment

Note: this schema will supersede page_edit, but we cannot remove it
right away as we first need to switch the extension and the http proxy
to use this one.

Bug: T124741
Change-Id: Iaa3044dbda30a45d2b38a6a8b4a22dbf0efc90f8