Include information about search results in CirrusSearchRequestSet
Adds CirrusSearchHit to the schema at both top level and per-request.
This represents what was returned to the user and at the per-request
level to identify what was returned from elasticsearch.

The .avsc file is buiilt from the idl. The idl format is much easier
to maintain from a human perspective. The .avsc is built as follows, the
avro-tools jar is a standard artifact built as part of apache/avro.

  java -jar avro-tools-1.7.7.jar idl2schemata CirrusSearchRequestSet.idl /tmp/csrq
  mv /tmp/csrq/CirrusSearchRequestSet.avsc 123456789.avsc

TODO: Would be best if jenkins could build the .avsc file for us

Bug: T128533
Change-Id: I1f194cc318c18e7c292c5450ea011811de8df204