Introduce triggeredBy field and move retry and continue schemas
To prevent infinite processing loops, where processing an event causes the
exact same event to be emitted, as well as for debugging purpose, we decided
to introduce an analogue of the X-Forwarded-For header to events where it makes
sence. Mediawiki events are not affected, because they always stay in the very
beginning of the processing chain.

I've decided to choose a comma-separated string for the triggeredBy field
instead of the array because for processing it will always be transformed to
the X-Triggered-By header, so to avoid joining and splitting strings on every
step, it's better to have it in the header format right away. Also, I don't see
any use-cases for a list.

Additionally, on the previous EventBus meeting we've decided to move the retry
and continue schemas to change-prop namespace, as right now they are specific
to change-prop, and we don't see any other use cases for them.

Change-Id: Icaa62a1cb6cfdb9f658c34ad514395a931698709