Add page, revision, and user entity based schemas for more consistent change modeling
Based on lots of discussion in https://gerrit.wikimedia.org/r/#/c/288210/
and with Dan, Aaron and Petr, We've reorganized the schemas to base them
on Mediawiki entities with consistent properties and prior state for change capture.

See jsonschema/mediawiki/README.md for more info.

Add simple comments for changes on this commit, but let's
discuss the larger idea in phabricator.

If this commit is accepted, we will abandon https://gerrit.wikimedia.org/r/#/c/288210/

This change also modifies jsonschema linting to only consider .yaml and .json files.

Bug: T134502
Change-Id: I0354461efc1b08bb5e737506fb2aa11e9ec9f270