Added sequence_num to the continue event.
For transclusions deduplication we need to compare
continue tokens (which will appear in the sequence
of continue tokens before the other one). But MW API
doesn't document how the tokens are sorted. To avoid
relying on the undocumented behaviour we cat add the
seq number to the event so that we can compare them.

Bug: T145011
Change-Id: I3eeea858cf5cb8a7ce84cf0141e4caaa996d7d31