Event schema for job queue event.
A very early proposal for a schema for the JobQueue event.
Mostly modelled after what's currently going on in the JobQueue,
with the addition of our `meta` parameter.

Potential improvement - generalise the deduplication information
pieces and reuse in other events, but too early for that.

Tested in Vagrant trying to do random stuff and sending these events,
seems that all of the jobs I've got in Vagrant actually conform to this
schema.

Bug: T157094
Change-Id: Ie83d0a8ba7d83c4b6e036c23c0eac877560e155a