build: Remove unused package.json properties
This isn't currently published or imported as an npm-package.
These properties were presumably filled in to avoid warnings
from npm, however, in this case it seems 'private' mode is more
appropiate, given it isn't actually meant to be a package currently
(it just uses npm for building/testing). The information here was
either false (e.g. non-existent "index.js"), duplicated from
elsewhere, or confusing.

The license claim tends to be a reflection of a LICENSE file
for the purpose of showing on npmjs.org pages. Given the claim
by itself isn't sufficient, I've just removed it, but if this
repo is meant to be Apache licensed, let's add a LICENSE file
to the repo.

Change-Id: I0159970baf532e5ede476d048e0439aa24cbc0e3