Increase date-time maxLength to 128 for all schemas
This is not a field or type change, so we do not bump schema versions.
maxLength is set to allow AJV to enforce safe schema validation.
(An unlimited length string validating against a date-format regex can cause
cpu load).

This allows us to accept longer date-times with microseconds and/or timezone
info.

Bug: T206789
Bug: T227896
Change-Id: Iec1706f1fc7d7e59023b17b4958ea11f4b6edab4